requirements.txt
altgraph==0.17.3
certifi==2020.6.20
dis3==0.1.3
Django==1.11.29
future==0.18.3
image==1.5.32
numpy==1.16.6
pefile==2019.4.18
Pillow==6.2.2
PyInstaller==3.6
PyOpenGL==3.1.5
pytz==2020.1
pywin32-ctypes==0.2.0
six==1.15.0
wincertstore==0.2
